Fix failing test for large collection instances.
